feat: Refactor for relocation support
This is the first step to support integration of this repo as a either
a submodule, a subtree, or as a snapshot copied to a target.

The functional code is largely unchanged, but dependencies and includes
have been updated to use `find_in_parent_folders` in preference to
`get_repo_root`.